Start of tradingcalendar commits.
Add a README file explaining the possible directions of this project.

Commits prior to this patch were filtered out of the Zipline project;
removing files that were not related to tradingcalendar logic.
This was done to extract just the trading calendar logic from Zipline,
while maintaining proper attribution.

Some commit messages may seem out of place, because the other files
touched were removed via filter, to which I apologize to the original
authors.

The current modules are not intended to be in a working or buildable
state, but to provide a base to start work on a standalone
tradingcalendar library which does not require the installation of
Zipline.